+++ packages/terra-functional-testing/tests/jest/commands/utils/index.test.js
@@ -6,11 +6,9 @@ const {
   downloadScreenshots,
   eventEmitter,
   getViewports,
-  getViewportSize,
   hideInputCaret,
   setApplicationLocale,
   setViewport,
-  setViewportSize,
 } = require('../../../../src/commands/utils');
 
 describe('index', () => {
@@ -22,10 +20,8 @@ describe('index', () => {
     expect(downloadScreenshots).toBeDefined();
     expect(eventEmitter).toBeDefined();
     expect(getViewports).toBeDefined();
-    expect(getViewportSize).toBeDefined();
     expect(hideInputCaret).toBeDefined();
     expect(setApplicationLocale).toBeDefined();
     expect(setViewport).toBeDefined();
-    expect(setViewportSize).toBeDefined();
   });
 });